preparing cart view
