--- src/css/App.css
@@ -1,37 +1,0 @@
-.container {
-  display: grid;
-  height: 100vh;
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr 6fr 1fr;
-  gap: 0px 0px;
-  grid-template-areas:
-    "header"
-    "recipe"
-    "footer";
-}
-
-.header {
-  grid-area: header;
-  display: flex;
-  justify-content: space-between;
-}
-
-
-
-.recipe {
-  grid-auto-rows: minmax(30px, auto);
-  grid-area: recipe;
-  display: flex;
-  margin: 5pt;
-  background-color: blueviolet;
-  border-radius: 20pt;
-  /*TODO is it needed ? 10/02/22*/
-}
-
-.footer {
-  grid-area: footer;
-  display: flex;
-  background-color: aqua;
-  justify-content: space-around;
-  align-items: center;
-}
