+++ src/css/Header.css
@@ -2,7 +2,7 @@
   margin-right: 20pt;
   margin-left: 20pt;
   margin-top: 10pt;
-  height: 50pt;
+  height: 30pt;
 }
